disjoint set of common divisors

+++ src/abc142/disjoint_set_of_common_divisors.h
@@ -1,1 +1,2 @@
-int disjoint_set_of_common_divisors(int a, int b);
+long long disjoint_set_of_common_divisors(long long a, long long b);
+
